feat(createTestingLibraryRule): add detectionOptions config
