+++ lib/create-testing-library-rule/index.ts
@@ -1,6 +1,9 @@
 import { ESLintUtils, TSESLint } from '@typescript-eslint/experimental-utils';
-import { getDocsUrl } from './utils';
+
+import { getDocsUrl } from '../utils';
+
 import {
+  DetectionOptions,
   detectTestingLibraryUtils,
   EnhancedRuleCreate,
 } from './detect-testing-library-utils';
@@ -11,24 +14,25 @@ type CreateRuleMeta<TMessageIds extends string> = {
   docs: CreateRuleMetaDocs;
 } & Omit<TSESLint.RuleMetaData<TMessageIds>, 'docs'>;
 
-export function createTestingLibraryRule<
+export const createTestingLibraryRule = <
   TOptions extends readonly unknown[],
   TMessageIds extends string,
   TRuleListener extends TSESLint.RuleListener = TSESLint.RuleListener
->(
-  config: Readonly<{
-    name: string;
-    meta: CreateRuleMeta<TMessageIds>;
-    defaultOptions: Readonly<TOptions>;
-    create: EnhancedRuleCreate<TOptions, TMessageIds, TRuleListener>;
-  }>
-): TSESLint.RuleModule<TMessageIds, TOptions> {
-  const { create, ...remainingConfig } = config;
-
-  return ESLintUtils.RuleCreator(getDocsUrl)({
+>({
+  create,
+  detectionOptions = {},
+  ...remainingConfig
+}: Readonly<{
+  name: string;
+  meta: CreateRuleMeta<TMessageIds>;
+  defaultOptions: Readonly<TOptions>;
+  detectionOptions?: Partial<DetectionOptions>;
+  create: EnhancedRuleCreate<TOptions, TMessageIds, TRuleListener>;
+}>): TSESLint.RuleModule<TMessageIds, TOptions> =>
+  ESLintUtils.RuleCreator(getDocsUrl)({
     ...remainingConfig,
     create: detectTestingLibraryUtils<TOptions, TMessageIds, TRuleListener>(
-      create
+      create,
+      detectionOptions
     ),
   });
-}
